Modify Script
Create Game Manager Script
(Infinite Stair Generation)
