+++ Platformgame-QSS-master (1)/Platformgame-QSS-master/Platformgame/Assets/Kimjimin/GameManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73a2ae2a64ac6b447a5d5e4492b24df4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
