+++ Platformgame-QSS-master (1)/Platformgame-QSS-master/Platformgame/Assets/Kimjimin/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40e69099ee394cc46bf8409d27b13309
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
