+++ Platformgame-QSS-master (1)/Platformgame-QSS-master/Platformgame/Assets/Kimjimin/Prefabs/Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: aa7bb16fa7ec46e4386f9567a08ee91a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
